backend/requirements-render.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
openpyxl==3.1.5
httpx==0.28.1
gspread==6.1.4
google-auth==2.38.0
google-api-python-client==2.164.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
